Addressed code review comments.
#28
